Use Padlayout with 3 Sektors per module. Correct calculation of hit position fromdigi info.
Correct calculation of hit position from dig info in class CbmTrdModule.
In the previous version the sector info was completely neglected and so the
result was wrong. Add a special case to initialize the data members for a
single sector in the constructor. Add new functions to do the coordinate
transformations. Before everything was in one function and not verry readable.
Add Comments to make calculations more clear.
In the class CbmTrdCreateDigiPar the default is now a layout with 3 pad
sectors per detector module to calculate the digitization parameters.

git-svn-id: https://subversion.gsi.de/fairroot/cbmroot/trunk@8802 0381ead4-6506-0410-b988-94b70fbc4730