CbmTrdModule.GetPosition returns now hit position and padsize.
The function GetPosition of the class CbmTrdModule now returns two TVector3.
One with the hit position in global coordinates as before, the second one
now contains the padsize of the fired pad. This information is now used when
calculating the error in CbmTrdHitProducerDigi and when drawing the digi in
CbmTrdDigiDraw. Add also new data member and Setters in CbmTrdDigiDraw to
switch on and of layers to be drawn.

git-svn-id: https://subversion.gsi.de/fairroot/cbmroot/trunk@8820 0381ead4-6506-0410-b988-94b70fbc4730